requirements.txt
numpy==1.15.0
pandas==0.23.3
python-dateutil==2.7.3
python-libsbml==5.17.0
pytz==2018.5
six==1.11.0
